tools/requirements.txt
feedparser>=6.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
google-generativeai>=0.3.0
firebase-admin>=6.2.0
